refactor(platform-ui): reuse form inputs
