+++ packages/platform-ui/src/pages/LLMModelUpsertPage.tsx
@@ -3,8 +3,8 @@ import { useForm } from 'react-hook-form';
 import { useNavigate, useParams } from 'react-router-dom';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import { Button } from '@/components/Button';
-import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
-import { Input } from '@/components/ui/input';
+import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/forms/Form';
+import { Input } from '@/components/Input';
 import { SelectInput } from '@/components/SelectInput';
 import { useLLMProviders } from '@/api/hooks/useLLMProviders';
 import { useCreateLLMModel, useLLMModel, useUpdateLLMModel } from '@/api/hooks/useLLMModels';
